[ 1755232 ] 3.1.2 tokenized program crashes on 3.2.0 revsion 590
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@600 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/classes/StringClass.hpp
@@ -303,7 +303,7 @@ class RexxStringClass : public RexxClass {
 
    inline size_t  getLength() { return this->length; };
    inline PCHAR getStringData() { return this->stringData; };
-   inline void put(size_t s, const void *b, int l) { memcpy((this->stringData+s),b,(size_t)l); };
+   inline void put(size_t s, const void *b, size_t l) { memcpy((this->stringData+s), b, l); };
    inline void put(size_t s, RexxString *o) { put(s, o->getStringData(), o->getLength()); };
    inline void set(size_t s,int c,int l) { memset((this->stringData+s),c,(size_t)l); };
    inline char getChar(size_t p) { return *(this->stringData+p); };
